--- src/compiler/typescript_to_rust/azle_generate/src/ts_ast/azle_type/azle_tuple_type/get_source_text.rs
@@ -1,8 +1,0 @@
-use super::AzleTupleType;
-use crate::ts_ast::{source_map::GetSourceFileInfo, GetSourceText};
-
-impl GetSourceText for AzleTupleType<'_> {
-    fn get_source_text(&self) -> String {
-        self.source_map.get_source(self.ts_tuple_type.span)
-    }
-}
