--- src/compiler/typescript_to_rust/azle_generate/src/ts_ast/azle_type/azle_tuple_type/to_data_type.rs
@@ -1,12 +1,0 @@
-use cdk_framework::act::node::{candid::Tuple, CandidType};
-
-use super::AzleTupleType;
-
-impl AzleTupleType<'_> {
-    pub fn to_data_type(&self) -> CandidType {
-        CandidType::Tuple(Tuple {
-            name: None,
-            elems: self.get_elem_types(),
-        })
-    }
-}
